+++ src/DigitalPreservation/DigitalPreservation.UI/Program.cs
@@ -95,10 +95,10 @@ try
         .AddSingleton<IMetsLoader, S3MetsLoader>()
         .AddSingleton<IMetsParser, MetsParser>()
         .AddSingleton<IMetsManager, MetsManager>()
-        .AddSingleton<IMetadataManager, MetadataManager>()
-        .AddSingleton<IPremisManager<FileFormatMetadata>, PremisManager>()
-        .AddSingleton<IPremisManager<ExifMetadata>, PremisManagerExif>()
-        .AddSingleton<IPremisEventManager<VirusScanMetadata>, PremisEventManagerVirus>()
+        .AddSingleton<MetadataManager>()
+        .AddSingleton<PremisManager>()
+        .AddSingleton<PremisManagerExif>()
+        .AddSingleton<PremisEventManagerVirus>()
         .AddSingleton<IMetsStorage, S3MetsStorage>()
         .AddSingleton<WorkspaceManagerFactory>()
         .AddCorrelationIdHeaderPropagation()
