+++ src/DigitalPreservation/Preservation.API/Program.cs
@@ -78,10 +78,10 @@ try
         .AddSingleton<IMetsParser, MetsParser>()
         .AddSingleton<IMetsManager, MetsManager>()
         .AddSingleton<IMetsFromArchivalGroup, MetsFromArchivalGroup>()
-        .AddSingleton<IMetadataManager, MetadataManager>()
-        .AddSingleton<IPremisManager<FileFormatMetadata>, PremisManager>()
-        .AddSingleton<IPremisManager<ExifMetadata>, PremisManagerExif>()
-        .AddSingleton<IPremisEventManager<VirusScanMetadata>, PremisEventManagerVirus>()
+        .AddSingleton<MetadataManager>()
+        .AddSingleton<PremisManager>()
+        .AddSingleton<PremisManagerExif>()
+        .AddSingleton<PremisEventManagerVirus>()
         .AddSingleton<IMetsStorage, S3MetsStorage>()
         .AddSingleton<WorkspaceManagerFactory>()
         .AddPreservationHealthChecks()
