+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerPathFixtureTests.cs
@@ -68,7 +68,7 @@ public class MetsManagerPathFixtureTests
         var premisManagerExif = new PremisManagerExif();
         var premisEventManager = new PremisEventManagerVirus();
         var metadataManager = new MetadataManager(premisManager, premisManagerExif, premisEventManager);
-        metsManager = new MetsManager(parser, metsStorage, metadataManager, premisManager, premisEventManager);
+        metsManager = new MetsManager(parser, metsStorage, metadataManager);
     }
 
     // -----------------------------------------------------------------------
