+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerPathTests.cs
@@ -61,7 +61,7 @@ public class MetsManagerPathTests
         var premisManagerExif = new PremisManagerExif();
         var premisEventManager = new PremisEventManagerVirus();
         var metadataManager = new MetadataManager(premisManager, premisManagerExif, premisEventManager);
-        metsManager = new MetsManager(parser, metsStorage, metadataManager, premisManager, premisEventManager);
+        metsManager = new MetsManager(parser, metsStorage, metadataManager);
     }
 
     // -----------------------------------------------------------------------
