+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerTests.cs
@@ -32,8 +32,8 @@ public class MetsManagerTests
         var premisManagerExif = new PremisManagerExif();
         var premisEventManager = new PremisEventManagerVirus();
         var metadataManager = new MetadataManager(premisManager, premisManagerExif, premisEventManager);
-        metsManager = new MetsManager(parser, metsStorage, metadataManager, premisManager, premisEventManager);
-        metsFromArchivalGroup = new MetsFromArchivalGroup(metsManager, parser);
+        metsManager = new MetsManager(parser, metsStorage, metadataManager);
+        metsFromArchivalGroup = new MetsFromArchivalGroup(metsManager, parser, metadataManager);
     }
 
     [Fact]
