+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerWithPremis.cs
@@ -53,7 +53,7 @@ public class MetsManagerWithPremis
         var premisManagerExif = new PremisManagerExif();
         var premisEventManager = new PremisEventManagerVirus();
         var metadataManager = new MetadataManager(premisManager, premisManagerExif, premisEventManager);
-        metsManager = new MetsManager(parser, metsStorage, metadataManager, premisManager, premisEventManager);
+        metsManager = new MetsManager(parser, metsStorage, metadataManager);
     }
 
     // -----------------------------------------------------------------------
